Got all the tests passing for pytest-shutil. Trimmed some of the duplicate and unnecessary functions.
